Split ht16k33 into separate components
