+++ components/ht16k33_base/README.md
@@ -1,0 +1,4 @@
+# HT16K33 4 character display base component
+
+This component provides the common functionality for both the 4 character 14 segment alphanumeric display and the 4 character (plus colon) 7 segment character display. It cannot be used as standalone. Use the ``ht16k33_alpha`` or ``ht16k33_7segment`` components instead.
+
