+++ components/ht16k33_base/ht16k33_display.cpp
@@ -2,23 +2,16 @@
 #include "esphome/core/hal.h"
 #include "esphome/core/helpers.h"
 #include "ht16k33_display.h"
-#include "font.h"
-
-#ifndef USE_ESP8266
-  #define pgm_read_word(s) (*s)
-#endif
 
 namespace esphome {
 namespace ht16k33 {
 
 static const char *TAG = "ht16k33";
 
-// First set bit determines command, bits after that are the data.
-static const uint8_t DISPLAY_COMMAND_SET_DDRAM_ADDR = 0x00;
-static const uint8_t DISPLAY_COMMAND_SYSTEM_SETUP = 0x21;
-static const uint8_t DISPLAY_COMMAND_DISPLAY_OFF = 0x80;
-static const uint8_t DISPLAY_COMMAND_DISPLAY_ON = 0x81;
-static const uint8_t DISPLAY_COMMAND_DIMMING = 0xE0;
+void HT16K33BaseDisplay::dump_config() {
+  ESP_LOGCONFIG(TAG, "HT16K33 Display:");
+  LOG_I2C_DEVICE(this);
+}
 
 void HT16K33BaseDisplay::setup() {
   for (auto *display : this->displays_) {
@@ -57,33 +50,6 @@ void HT16K33BaseDisplay::loop() {
 
 float HT16K33BaseDisplay::get_setup_priority() const { return setup_priority::PROCESSOR; }
 
-void HT16K337SegmentDisplay::display_() {
-  int offset = this->offset_;
-  constexpr uint8_t size = 10;
-  uint8_t buffer[size];
-  memcpy(buffer, this->buffer_ + offset, 4);
-  if (this->show_colon_) {
-    buffer[4] = 0x02;
-  } else {
-    buffer[4] = 0;
-  }
-  buffer[5] = 0;
-  memcpy(buffer + 6, this->buffer_ + offset + 4, 4);
-
-  for (auto *display : this->displays_) {
-    display->write_bytes(DISPLAY_COMMAND_SET_DDRAM_ADDR, buffer, size);
-    offset += 8;
-  }
-}
-
-void HT16K33AlphaDisplay::display_() {
-  int offset = this->offset_;
-  for (auto *display : this->displays_) {
-    display->write_bytes(DISPLAY_COMMAND_SET_DDRAM_ADDR, this->buffer_ + offset, 8);
-    offset += 8;
-  }
-}
-
 void HT16K33BaseDisplay::update() {
   memset(this->buffer_, 0, 64);
   int prev_fill = this->buffer_fill_;
@@ -119,7 +85,9 @@ float HT16K33BaseDisplay::get_brightness() {
 
 void HT16K33BaseDisplay::print(const char *str) {
   uint8_t pos = this->buffer_fill_;
+  uint8_t idx = 1;
   uint16_t fontc = 0;
+  this->show_colon_ = false;
   while (*str != '\0') {
     if (pos >= 64) {
       ESP_LOGW(TAG, "output buffer full!");
@@ -135,9 +103,16 @@ void HT16K33BaseDisplay::print(const char *str) {
     if (c == '.') {
       fontc |= decimal_point_mask_();
       str++;
+      idx++;
+    }
+    if (c == ':' && idx == 2 && this->supports_colon_()) {
+      this->show_colon_ = true;
+      str++;
+      idx++;
     }
     this->buffer_[pos++] = fontc & 0xff;
     this->buffer_[pos++] = fontc >> 8;
+    idx++;
   }
   this->buffer_fill_ = pos;
 }
@@ -162,14 +137,6 @@ void HT16K33BaseDisplay::strftime(const char *format, ESPTime time) {
     this->print(buffer);
 }
 #endif
-
-uint16_t HT16K33AlphaDisplay::read_character_(uint8_t c) const {
-  return pgm_read_word(&alphafonttable[c]);
-}
-
-uint16_t HT16K337SegmentDisplay::read_character_(uint8_t c) const {
-  return pgm_read_word(&sevensegfonttable[c - 32]);
-}
 
 }  // namespace ht16k33
 }  // namespace esphome
